메인홈-베스트,신상품,간식 공통함수 작업
