+++ src/components/Home/PrdSnack.jsx
@@ -1,48 +1,28 @@
-import React, { useState, useEffect } from 'react';
-import axios from 'axios';
+import React,{useEffect} from 'react';
 
 // Compoennts
 import ProductItem from "../Card/ProductItem";
+import useProductRandom from '../../hooks/useProductRandom';
 
 //css
 import styles from "../../styles/css/components/Category/Snack.module.css";
 
 //메인 홈 - 스낵 베너
 const PrdSnack = (props) => {
-   const [snack, setSnack] = useState([]);
-  // const {id} = useParams();
-
+  const url = "/products/categories?categories=SNACK";
+  const [random, getRandom ] = useProductRandom(url);
+  
   useEffect(() => {
-    const url = `http://ec2-15-164-206-172.ap-northeast-2.compute.amazonaws.com/products/categories?categories=SNACK`;
-    const data = async () => {
-      try {
-        const res = await axios.get(url);
-        console.log(res);
-        const randomSnackItem = [];
-        while(true){
-          if(randomSnackItem.length==4){
-            break;
-          }
-          const random = Math.floor(Math.random() * res.data.length);
-          if(randomSnackItem.filter(item => item.id===random + 1).length ===0){
-            randomSnackItem.push(res.data[random]);
-          }
-        }
-        // console.log(randomBestItem);
-         setSnack(randomSnackItem);
-      }catch(error) {
-        console.log(error);
-      }
-    }
-    data();
+    getRandom();
   },[]);
+
   return (
       <div className={styles.snack_Slide}>
         <div className={styles.snack_Slide_hea}>
           <h2>영양 만점 간식 추천</h2>
         </div>
         <div className={styles.snack_Slide_prd}>
-          <ProductItem data={snack} className='container-4columns'/>
+          <ProductItem data={random} className='container-4columns'/>
         </div>
       </div>
   );
